requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
streamlit>=1.20.0
Pillow>=8.3.0
scikit-learn>=1.0.0